### 2024-10-22 #7395 @editorialbot
[PRE REVIEW]: antimeridian: A Python package for correcting geometries that cross the 180th meridian
**Submitting author:** @gadomski (<a href="http://orcid.org/0000-0003-4877-7217">Peter Gadomski</a>)
**Repository:** https://github.com/gadomski/antimeridian
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.9
