### 2024-10-24 #7395 @kthyng
Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file.

I just wish your package handled every problem with longitude for every user! 😆
