### 2024-10-24 e089465
docs: fix capitalization in bib file
