+++ docs/paper.bib
@@ -1,6 +1,6 @@
 @rfc{10.17487/RFC7946,
 author = {Butler, H. and Daly, M. and Doyle, A. and Gillies, S. and Hagen, S. and Schaub, T.},
-title = {RFC 7946: The GeoJSON Format},
+title = {{RFC 7946: The GeoJSON Format}},
 year = {2016},
 publisher = {RFC Editor},
 address = {USA},
@@ -9,7 +9,7 @@ abstract = {GeoJSON is a geospatial data interchange format based on JavaScript 
 
 @proceedings{alma992356353405961,
 author = {Various},
-title = {International Conference Held at Washington for the Purpose of Fixing a Prime Meridian and a Universal Day. October, 1884. Protocols of the Proceedings},
+title = {{International Conference Held at Washington for the Purpose of Fixing a Prime Meridian and a Universal Day. October, 1884. Protocols of the Proceedings}},
 year = {1884},
 url = {https://www.gutenberg.org/files/17759/17759-h/17759-h.htm}
 }
@@ -34,7 +34,7 @@ year = {2024}
 
 @manual{Cartopy,
 author = {{Met Office}},
-title = {Cartopy: a cartographic python library with a Matplotlib interface},
+title = {{Cartopy: a cartographic python library with a Matplotlib interface}},
 year = {2010 - 2015},
 address = {Exeter, Devon },
 url = {https://scitools.org.uk/cartopy}
