### 2024-10-24 #7395 @gadomski
> Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file.

Should be fixed by https://github.com/gadomski/antimeridian/pull/152, thanks for the catch! TIL.

> I just wish your package handled every problem with longitude for every user! 😆

Right?! We do what we can but it's a messy world out there.
