### 2024-11-26 #7530 @editorialbot
[REVIEW]: antimeridian: A Python package for correcting geometries that cross the 180th meridian
**Submitting author:** @gadomski (<a href="http://orcid.org/0000-0003-4877-7217">Peter Gadomski</a>)
**Repository:** https://github.com/gadomski/antimeridian
**Branch with paper.md** (empty if default branch):
**Version:** 0.3.12
