### 2024-11-26 #7530 @mikemahoney218
👋🏼 @gadomski, @ianturton, @busstoptaktik: this is the review thread for the paper. Just about all of our communications will happen here from now on 😄

```@editorialbot generate my checklist```

We aim for reviews to be completed within about 4 weeks. Please let me know if you require some more time!
