### 2024-11-26 #7530 @ianturton
I would like to see more details of the differences between the output of this package and GDAL's `--wrap-dateline` , as this is currently state of the art and potential users may be confused by unexpected differences between the two.
