### 2024-11-26 gadomski/antimeridian#155 @gadomski
Write up differences between `ogr2ogr --wrap-dateline`
Both as a docs page and in the paper. Requested in https://github.com/openjournals/joss-reviews/issues/7530#issuecomment-2500609228
