### 2024-11-26 #7530 @gadomski
I've got the `ogr2ogr --wrap-dateline` output in https://github.com/gadomski/antimeridian/tree/main/tests/data/ogr2ogr, so would just need to write it up into a docs page and then include a snippet in the paper. https://github.com/gadomski/antimeridian/issues/155
