### 2024-12-01 7a60e7d
feat: switch to mkdocs-material
