+++ src/antimeridian/_implementation.py
@@ -55,12 +55,12 @@ class FixWindingWarning(AntimeridianWarning):
 
 
 class GeoInterface(Protocol):
-    """A simple protocol for things that have a ``__geo_interface__`` method.
+    """A simple protocol for things that have a `__geo_interface__` method.
 
-    The ``__geo_interface__`` protocol is described `here
-    <https://gist.github.com/sgillies/2217756>`_, and is used within `shapely
-    <https://shapely.readthedocs.io/en/stable/manual.html>`_ to
-    extract geometries from objects.
+    The `__geo_interface__` protocol is described
+    [here](https://gist.github.com/sgillies/2217756>), and is used within
+    [shapely](https://shapely.readthedocs.io/en/stable/manual.html) to extract
+    geometries from objects.
     """
 
     @property
@@ -78,8 +78,8 @@ def fix_geojson(
 
     If the object does not cross the antimeridian, it is returned unchanged.
 
-    See :py:func:`fix_polygon` for a description of the ``force_north_pole``,
-    ``force_south_pole``, and ``fix_winding`` arguments.
+    See [antimeridian.fix_polygon][] for a description of the `force_north_pole`
+    `force_south_pole` and `fix_winding` arguments.
 
     Args:
         geojson: A GeoJSON object as a dictionary
@@ -170,13 +170,13 @@ def fix_shape(
 ) -> Dict[str, Any]:
     """Fixes a shape that crosses the antimeridian.
 
-    See :py:func:`fix_polygon` for a description of the ``force_north_pole``,
-    ``force_south_pole``, and ``fix_winding`` arguments.
+    See [antimeridian.fix_polygon][] for a description of the `force_north_pole`
+    `force_south_pole` and `fix_winding` arguments.
 
     Args:
         shape: A polygon, multi-polygon, line string, or multi-line string,
-            either as a dictionary or as a :py:class:`GeoInterface`. Uses
-            :py:func:`shapely.geometry.shape` under the hood.
+            either as a dictionary or as a [antimeridian.GeoInterface][]. Uses
+            [shapely.geometry.shape][] under the hood.
         force_north_pole: If the polygon crosses the antimeridian, force the
             joined segments to enclose the north pole.
         force_south_pole: If the polygon crosses the antimeridian, force the
@@ -242,10 +242,10 @@ def fix_multi_polygon(
     force_south_pole: bool = False,
     fix_winding: bool = True,
 ) -> MultiPolygon:
-    """Fixes a :py:class:`shapely.geometry.MultiPolygon`.
+    """Fixes a [shapely.MultiPolygon][].
 
-    See :py:func:`fix_polygon` for a description of the ``force_north_pole``,
-    ``force_south_pole``, and ``fix_winding`` arguments.
+    See [antimeridian.fix_polygon][] for a description of the `force_north_pole`
+    `force_south_pole` and `fix_winding` arguments.
 
     Args:
         multi_polygon: The multi-polygon
@@ -277,22 +277,22 @@ def fix_polygon(
     force_south_pole: bool = False,
     fix_winding: bool = True,
 ) -> Union[Polygon, MultiPolygon]:
-    """Fixes a :py:class:`shapely.geometry.Polygon`.
+    """Fixes a [shapely.Polygon][].
 
     If the input polygon is wound clockwise, it will be fixed to be wound
-    counterclockwise _unless_ ``fix_winding`` is ``False``, in which case it
+    counterclockwise _unless_ `fix_winding` is `False` in which case it
     will be corrected by adding a counterclockwise polygon from (-180, -90) to
     (180, 90) as its exterior.
 
     In rare cases, the underlying algorithm might need a little help to fix the polygon.
[…]
