+++ docs/api.rst
@@ -1,5 +1,0 @@
-API documentation
-=================
-
-.. automodule:: antimeridian
-    :members:
