### 2024-12-01 d452194
feat: add docs to ci
