+++ .github/workflows/ci.yaml
@@ -12,6 +12,7 @@ on:
 
 jobs:
   test:
+    name: Test
     runs-on: ubuntu-latest
     strategy:
       fail-fast: true
@@ -33,6 +34,7 @@ jobs:
       - name: pytest
         run: uv run pytest
   min-dependencies:
+    name: Minimum dependencies
     runs-on: ubuntu-latest
     steps:
       - name: Checkout
@@ -50,25 +52,3 @@ jobs:
         run: uv sync --resolution lowest-direct --all-extras
       - name: pytest
         run: uv run pytest
-  docs:
-    runs-on: ubuntu-latest
-    defaults:
-      run:
-        shell: bash -el {0}
-    env:
-      PYDEVD_DISABLE_FILE_VALIDATION: 1
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-      - name: Cache conda
-        uses: actions/cache@v4
-        with:
-          path: ~/conda_pkgs_dir
-          key: ${{ runner.os }}-conda-${{ hashFiles('docs/environment.yaml') }}
-      - uses: conda-incubator/setup-miniconda@v3
-        with:
-          auto-update-conda: true
-          environment-file: docs/environment.yaml
-          use-mamba: true
-      - name: Check docs
-        run: sphinx-build -W docs docs/_build/html
