### 2024-12-01 3270178
refactor: build docs on every main push
