+++ .github/workflows/docs.yaml
@@ -2,8 +2,8 @@ name: Docs
 
 on:
   push:
-    tags:
-      - v*
+    branches:
+      - main
   workflow_dispatch:
 
 permissions:
