### 2024-12-01 c1190a6
feat: switch docs to mkdocs-material (#156)
