+++ docs/the-algorithm.md
@@ -1,58 +1,46 @@
-The algorithm
-=============
+# The algorithm
 
 What follows is a walkthrough, with visual aids, of the algorithm underlying this package.
 
-Background
-~~~~~~~~~~
+## Background
 
 Let's start with a simple GeoJSON geometry that doesn't come anywhere near the antimeridian.
 For the sake of example, let's say it runs from 40° longitude west to 40° east, and from 20° latitude south to 20° north.
 Its lower left point is at ``-40, -20``, and its upper right point is at ``40, 20``.
 
-.. figure:: img/box.png
-    :scale: 30%
-    :align: center
-
-    A simple box.
+![A simple box](img/box.png)
 
 Now let's consider the same geometry, but shifted east 180 degrees.
 Its lower-left point is ``140, -20``, and its upper-right point is at ``220, 20``.
 Here's its representation in GeoJSON:
 
-.. literalinclude:: examples/crossing.json
+```json
+--8<-- "docs/examples/crossing.json"
+```
 
 And visualized:
 
-
-.. figure:: img/crossing.png
-    :scale: 30%
-    :align: center
-
-    A box that crosses the antimeridian.
+![Crossing](img/crossing.png)
 
 So far, good enough.
-However, the `bounds of WGS84 <https://epsg.io/4326>`_, the coordinate system used by GeoJSON, go from -180° to 180° longitude.
+However, the bounds of [WGS84](https://epsg.io/4326), the coordinate system used by GeoJSON, go from -180° to 180° longitude.
 This means that the upper-right coordinate of our antimeridian-crossing box should be "wrapped" to ``-140, 20``:
 
-.. literalinclude:: examples/wrapped.json
-    :diff: examples/crossing.json
+```json
+--8<-- "docs/examples/wrapped.json"
+```
 
 Let's consider the four points of our "wrapped" box, and the order in which they're connected:
 
-.. code-block:: text
+```text
+    1              2              3            4            1
 
-        1              2              3            4            1
-
-    [140, -20] -> [-140, -20] -> [-140, 20] -> [140, 20] -> [140, 20]
+[140, -20] -> [-140, -20] -> [-140, 20] -> [140, 20] -> [140, 20]
+```
 
 Exterior rings in GeoJSON should be wound counter-clockwise, but you can see that our antimeridian-crossing box appears to be wound clockwise.
 
-.. figure:: img/winding.png
-    :scale: 30%
-    :align: center
-
-    The four points in our box, with their connections.
+![Winding](img/winding.png)
 
 This leads to a confusing situation — because it's invalid GeoJSON!
 You should not have a standalone clockwise ring.
@@ -60,120 +48,69 @@ When rendering this polygon, some systems automatically correct the winding orde
 Another interpretation is that the shape is a hole, and the shell is the entire rest of the globe.
 This is very rarely what you're actually trying to represent.
 
-.. figure:: img/winding-repercussions.png
-    :scale: 30%
-    :align: center
-
-    How should we interpret this invalid GeoJSON?
+![Winding repercussions](img/winding-repercussions.png)
 
 Our algorithm can take this invalid GeoJSON as input and produce valid GeoJSON as output.
 
-Segmentation
-~~~~~~~~~~~~
+## Segmentation
 
 To correct the invalid GeoJSON, we break our polygon into **segments**.
 To create segments, we start at the first point of the geometry, and walk through the points until we find a longitude jump of more than 180° degrees.
 
-.. figure:: img/delta-gt-180.png
-    :scale: 30%
-    :align: center
-
-    A longitude jump greater than 180°.
+![Large delta](img/delta-gt-180.png)
 
 When we find such a jump, we split the polygon into segments by inserting two points, one on each antimeridian.
 The first point ends the first segment, and the second point starts the second segment.
 
[…]
