+++ pyproject.toml
@@ -57,6 +57,13 @@ dev-dependencies = [
     "pytest-console-scripts>=1.4",
     "pytest>=8.0",
     "ruff>=0.6.1",
+    "mkdocs-material[imaging]>=9.5.47",
+    "mkdocs-jupyter>=0.25.1",
+    "shapely>=2.0.6",
+    "scipy>=1.14.1",
+    "mkdocstrings[python]>=0.27.0",
+    "mkdocs-click>=0.8.1",
+    "mike>=2.1.3",
 ]
 
 [build-system]
