### 2024-12-01 a8094d5
docs: add redirect for readthedocs
