+++ pyproject.toml
@@ -64,6 +64,7 @@ dev-dependencies = [
     "mkdocstrings[python]>=0.27.0",
     "mkdocs-click>=0.8.1",
     "mike>=2.1.3",
+    "mkdocs-redirects>=1.2.2",
 ]
 
 [build-system]
