+++ mkdocs.yml
@@ -63,6 +63,10 @@ plugins:
             signature_crossrefs: true
           import:
             - https://shapely.readthedocs.io/en/stable/objects.inv
+  - redirects:
+      redirect_maps:
+        "en/stable/index.md": "index.md"
+        "en/latest/index.md": "index.md"
 
 markdown_extensions:
   - pymdownx.snippets
