### 2024-12-01 cada1e8
docs: syntax highlighting
