### 2024-12-01 4ca1c5f
fix: remove redirects
