### 2024-12-02 309bb2c
feat: add note to paper re: WWRAPDATELINE
