### 2024-12-02 #7530 @gadomski
@ianturton added some [text to the paper](https://github.com/gadomski/antimeridian/blob/309bb2c32da97fe69aab801e4e315130464b8658/docs/paper.md?plain=1#L66-L70) and a companion [comparison notebook](https://www.gadom.ski/antimeridian/latest/comparison/). Tl;dr: they looks like they produce the same output as long as you tune `-wrapdatelin` with `-datelineoffset`.
