### 2024-12-03 #7530 @mikemahoney218
While we can accept papers with two reviews, we always prefer to have three where possible -- so thank you @mmann1123 for agreeing to review this submission as well. Instructions for how to carry out a review are in this comment here:

https://github.com/openjournals/joss-reviews/issues/7530#issuecomment-2500343550

Thanks again!
