### 2024-12-03 0d55b1d
docs: add JOSS badge
