### 2024-12-03 #7530 @mmann1123
I think everything looks good, it was interesting to see the comparison to `ogr2ogr --wrap-dateline`.  Everything seems in order, installs easily, has relevant unit tests, the documentation is unusually good, and the paper is clear.
