### 2024-12-07 #7530 @busstoptaktik
I fully agree with @mmann1123's characterization in https://github.com/openjournals/joss-reviews/issues/7530#issuecomment-2515023093 - so please hit the publish button!
