### 2024-12-07 #7530 @mikemahoney218
Haha -- thank you so much, @busstoptaktik , and congratulations @gadomski on three glowing reviews.

@gadomski , the next step is for you to work through the "Author Tasks" section in the comment above to create a release and archive of the software. Once you've done that (and linked them here), I'll go through my section and move to recommend acceptance.

As a quick note, I'll be at AGU this coming week and so might be a bit longer in responding to this thread -- apologies in advance!
