+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "antimeridian"
-version = "0.3.11"
+version = "0.3.12"
 authors = [{ name = "Pete Gadomski", email = "pete.gadomski@gmail.com" }]
 description = "Fix GeoJSON geometries that cross the 180th meridian"
 readme = "README.md"
