+++ .pre-commit-config.yaml
@@ -6,14 +6,14 @@ repos:
       - id: end-of-file-fixer
       - id: check-yaml
   - repo: https://github.com/pre-commit/mirrors-mypy
-    rev: v1.12.0
+    rev: v1.13.0
     hooks:
       - id: mypy
         additional_dependencies:
           - click~=8.1.6
           - pytest>=8.0
   - repo: https://github.com/charliermarsh/ruff-pre-commit
-    rev: v0.6.9
+    rev: v0.8.2
     hooks:
       - id: ruff
         types_or: [python, pyi, jupyter]
