### 2024-12-09 93ff394
docs: add doi badge
