### 2024-12-09 #7530 @gadomski
@mikemahoney218 here's my author tasks:

- [x] Make a release of the software with the latest changes from the review: [0.3.12](https://pypi.org/project/antimeridian/0.3.12/)
- [x] DOI: [10.5281/zenodo.14335984](https://doi.org/10.5281/zenodo.14335984)
- [x] Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper
