### 2024-12-10 d3c9e7e
docs: add conda badge
