### 2024-12-11 5097fbd
feat: add great-circle meridian crossing option
This commit adds an option to compute the intersection of a segment and
the meridian using spherical geometry, which provides nicer splits for
large crossing the meridian diagonally.

Related to #153
