+++ src/antimeridian/_implementation.py
@@ -74,6 +74,7 @@ def fix_geojson(
     force_north_pole: bool = False,
     force_south_pole: bool = False,
     fix_winding: bool = True,
+    great_circle: bool = False,
 ) -> Dict[str, Any]:
     """Fixes a GeoJSON object that crosses the antimeridian.
 
@@ -90,6 +91,8 @@ def fix_geojson(
             joined segments to enclose the south pole.
         fix_winding: If the polygon is wound clockwise, reverse its
             coordinates before applying the algorithm.
+        great_circle: Compute meridian crossings on the sphere rather than
+            using 2D geometry.
 
     Return:
         The same GeoJSON with a fixed geometry or geometries
@@ -106,6 +109,7 @@ def fix_geojson(
             force_north_pole=force_north_pole,
             force_south_pole=force_south_pole,
             fix_winding=fix_winding,
+            great_circle=great_circle,
         )
         return geojson
     elif type_ == "FeatureCollection":
@@ -118,6 +122,7 @@ def fix_geojson(
                 force_north_pole=force_north_pole,
                 force_south_pole=force_south_pole,
                 fix_winding=fix_winding,
+                great_circle=great_circle,
             )
         geojson["features"] = features
         return geojson
@@ -127,10 +132,11 @@ def fix_geojson(
             force_north_pole=force_north_pole,
             force_south_pole=force_south_pole,
             fix_winding=fix_winding,
+            great_circle=great_circle,
         )
 
 
-def segment_geojson(geojson: Dict[str, Any]) -> MultiLineString:
+def segment_geojson(geojson: Dict[str, Any], great_circle: bool) -> MultiLineString:
     """Segments a GeoJSON object into a MultiLineString.
 
     If the object does not cross the antimeridian, its exterior and interior
@@ -138,6 +144,8 @@ def segment_geojson(geojson: Dict[str, Any]) -> MultiLineString:
 
     Args:
         geojson: A GeoJSON object as a dictionary
+        great_circle: Compute meridian crossings on the sphere rather than
+            using 2D geometry.
 
     Return:
         A MutliLineString of segments.
@@ -149,17 +157,17 @@ def segment_geojson(geojson: Dict[str, Any]) -> MultiLineString:
         geometry = geojson.get("geometry", None)
         if geometry is None:
             raise ValueError("no 'geometry' field found in GeoJSON Feature")
-        return MultiLineString(segment_shape(geometry))
+        return MultiLineString(segment_shape(geometry, great_circle))
     elif type_ == "FeatureCollection":
         features = geojson.get("features", None)
         if features is None:
             raise ValueError("no 'features' field found in GeoJSON FeatureCollection")
         segments = list()
         for feature in features:
-            segments.extend(segment_geojson(feature))
+            segments.extend(segment_geojson(feature, great_circle))
         return MultiLineString(segments)
     else:
-        return MultiLineString(segment_shape(geojson))
+        return MultiLineString(segment_shape(geojson, great_circle))
 
 
 def fix_shape(
@@ -168,6 +176,7 @@ def fix_shape(
     force_north_pole: bool = False,
     force_south_pole: bool = False,
     fix_winding: bool = True,
+    great_circle: bool = False,
 ) -> Dict[str, Any]:
     """Fixes a shape that crosses the antimeridian.
 
@@ -184,6 +193,8 @@ def fix_shape(
             joined segments to enclose the south pole.
         fix_winding: If the polygon is wound clockwise, reverse its
             coordinates before applying the algorithm.
+        great_circle: Compute meridian crossings on the sphere rather than
+            using 2D geometry.
 
     Returns:
         The fixed shape as a dictionary
@@ -198,6 +209,7 @@ def fix_shape(
[…]
