### 2024-12-12 c42e2c5
feat: parametrize tests over great_circle values
