+++ tests/test_geojson.py
@@ -1,18 +1,22 @@
 import copy
 
+import pytest
+
 import antimeridian
 
 from .conftest import Reader
 
 
-def test_fix_feature(read_input: Reader) -> None:
+@pytest.mark.parametrize("great_circle", [True, False])
+def test_fix_feature(read_input: Reader, great_circle: bool) -> None:
     input = read_input("simple")
     feature = {"type": "Feature", "geometry": input, "properties": {"foo": "bar"}}
-    fixed = antimeridian.fix_geojson(feature)
+    fixed = antimeridian.fix_geojson(feature, great_circle=great_circle)
     assert fixed["properties"]["foo"] == "bar"
 
 
-def test_fix_feature_collection(read_input: Reader) -> None:
+@pytest.mark.parametrize("great_circle", [True, False])
+def test_fix_feature_collection(read_input: Reader, great_circle: bool) -> None:
     input = read_input("simple")
     feature_a = {
         "type": "Feature",
@@ -25,14 +29,15 @@ def test_fix_feature_collection(read_input: Reader) -> None:
         "features": [feature_a, feature_b],
         "another": "property",
     }
-    fixed = antimeridian.fix_geojson(feature_collection)
+    fixed = antimeridian.fix_geojson(feature_collection, great_circle=great_circle)
     assert fixed["features"][0]["properties"]["foo"] == "bar"
     assert fixed["features"][1]["properties"]["baz"] == "boz"
     assert fixed["another"] == "property"
 
 
-def test_segment_feature(read_input: Reader) -> None:
+@pytest.mark.parametrize("great_circle", [True, False])
+def test_segment_feature(read_input: Reader, great_circle: bool) -> None:
     input = read_input("split")
     feature = {"type": "Feature", "geometry": input}
-    fixed = antimeridian.segment_geojson(feature)
+    fixed = antimeridian.segment_geojson(feature, great_circle)
     assert len(fixed.geoms) == 2
