### 2024-12-12 9a2a47a
feat: move output test data to subdirectory
This prepares the addition of test outputs expected from spherical
geometry.
