### 2024-12-12 80a35db
feat: refactor Reader type alias as protocol
This is required to introduce an optional argument with a default value
(subdirectory), which cannot be expressed with Callable type hinting
only.

This has the nice benefit of not removing the need for typing
extensions, since Protocols are available since Python 3.8, which is
lower than the oldest non-EOL Python version.

See https://docs.python.org/3/library/typing.html#annotating-callable-objects
