### 2024-12-12 110402a
test: add great-circle polygon fixing test
The output of ogr2ogr is consistent with the behaviour of great-circle
meridian crossing computation introduced by this PR.
