### 2024-12-12 0cf6da7
docs: add entry to changelog
