### 2024-12-12 e4e47da
test: add great circle test suite
The differences have been checked visually and are convincing, except
the force-north-pole test for which a double-check would be needed.
