### 2024-12-12 de2401b
feat!: change default value of great_circle arg
Some tests were kept unparametrized over great_circle to make sure that
the default behaviour is also tested.

See https://github.com/gadomski/antimeridian/pull/159#issuecomment-2536808756

BREAKING CHANGE: Meridian crossings are now computed using spherical
geometry by default.
