### 2024-12-18 a054124
feat: add great-circle meridian crossing option (#159)
* feat: add great-circle meridian crossing option

This commit adds an option to compute the intersection of a segment and
the meridian using spherical geometry, which provides nicer splits for
large crossing the meridian diagonally.

Related to #153

* feat: add great_circle argument to functions

This propagates the decision of whether to use planar or spherical
geometry to compute meridian crossing latitudes.
The default is set to False to keep this commit self-contained, because
changing the default has a significant impact on the test
[…]
