### 2024-12-18 8a480b3
release: v0.4.0 (#160)
