+++ .pre-commit-config.yaml
@@ -13,7 +13,7 @@ repos:
           - click~=8.1.6
           - pytest>=8.0
   - repo: https://github.com/charliermarsh/ruff-pre-commit
-    rev: v0.8.2
+    rev: v0.8.3
     hooks:
       - id: ruff
         types_or: [python, pyi, jupyter]
