### 2024-12-24 #7530 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
