### 2024-12-31 #7530 @kthyng
@gadomski It looks like the DOI you provided is the general DOI for the package, but we should use the DOI specific to the version associated with your JOSS review, which looks like 10.5281/zenodo.14335985.
