### 2024-12-31 #7530 @kthyng
@gadomski Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. "python" should be capitalized.
