### 2024-12-31 #7530 @gadomski
@kthyng no problem ... however, I did copy that text directly from https://scitools.org.uk/cartopy/docs/latest/citation.html#bibtex-entry, which does _not_ capitalize "python"?
