### 2025-01-02 b00e3ff
feat: update citation capitalization
