### 2025-01-02 #7530 @gadomski
@kthyng done in https://github.com/gadomski/antimeridian/pull/162, do you need a new release or can we publish from **main**?

![Screenshot 2025-01-02 at 9 04 22 AM](https://github.com/user-attachments/assets/0b5a6ad6-b733-4be9-b3b0-99e6a130e5f3)
