### 2025-01-08 #7530 @kthyng
@gadomski Ah that is a good point. I would be fine with either then — I can't think of a better argument for one vs. the other though maybe we have to use their own citation?

Let me know what you want to do and to answer your question, no you don't need to do a new release since we're just changing the paper.
