### 2025-01-08 4ec8fa1
Revert "feat: update citation capitalization (#162)"
This reverts commit 993d1730fdd18ae186e15f68e0138a354dc5609b.
