### 2025-01-08 e722d12
Deployed 4ec8fa1 to v0.4.0 with MkDocs 1.6.1 and mike 2.1.3
