+++ v0.4.0/paper.bib
@@ -34,7 +34,7 @@ year = {2024}
 
 @manual{Cartopy,
 author = {{Met Office}},
-title = {{Cartopy: a cartographic Python library with a Matplotlib interface}},
+title = {{Cartopy: a cartographic python library with a Matplotlib interface}},
 year = {2010 - 2015},
 address = {Exeter, Devon },
 url = {https://scitools.org.uk/cartopy}
