### 2025-01-08 #7530 @gadomski
I have a light preference for following their recommendation, so I've reverted in https://github.com/gadomski/antimeridian/commit/4ec8fa135674044f3099592108a8ff42bb0f2c5b. I'm 👍🏼 to :shipit:.

![Screenshot 2025-01-08 at 1 24 14 PM](https://github.com/user-attachments/assets/20a3841f-3de4-4477-9c59-c54e91f268aa)
