### 2025-01-08 #7530 @kthyng
Congratulations on your new publication @gadomski! Many thanks to editor @mikemahoney218 and to reviewers @ianturton, @busstoptaktik, and @mmann1123 for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
